requirements.txt
aiogram==3.3.0
python-dotenv>=1.0.1
openai>=1.40.0
requests>=2.31.0
yookassa>=3.0.1

